Fixing the Navigator
