+++ app/src/main/java/com/potus/potus_front/MainActivity.kt
@@ -53,16 +53,13 @@ class MainActivity : ComponentActivity() {
         tokenStateViewModel.myLocation(getLastLocation())
 
         setContent {
-
             Potus_frontTheme {
                 // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
-                    // color = MaterialTheme.colors.background
                     color = MaterialTheme.colors.background
 
                 ) {
-
                     CompositionLocalProvider(TokenState provides tokenStateViewModel) {
                         Navigation()
                     }
