--- app/src/main/java/com/potus/potus_front/ui/screens/Pantalla1.kt
@@ -1,42 +1,0 @@
-package com.potus.potus_front.ui.screens
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.Button
-import androidx.compose.material.Text
-import androidx.compose.material.TextField
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun Pantalla1(){
-    var text by remember{
-        mutableStateOf("")
-    }
-    Column(
-        verticalArrangement = Arrangement.Center,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 50.dp)
-    ){
-        TextField(
-            value =text,
-            onValueChange = {
-                text = it
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        Button(
-            onClick = {
-
-            },
-            modifier = Modifier.align(Alignment.End)
-        ){
-            Text("ES LA PANTALLA 1")
-        }
-
-    }
-
-}
