+++ app/src/main/java/com/potus/potus_front/ui/screens/Screen.kt
@@ -1,15 +1,11 @@
 package com.potus.potus_front.ui.screens
 
 sealed class Screen(val route: String){
-    object MainScreen : Screen("main_screen")
-    object DetailScreen : Screen("detail_screen")
     object AuthScreen : Screen("auth_screen")
     object ProfileScreen : Screen("profile_screen")
     object HomeScreen : Screen("home_screen")
     object RegisterScreen : Screen("register_screen")
     object SwitcherScreen : Screen("switcher_screen")
-    object Pantalla1 : Screen("pantalla1")
-
 
     fun withArgs(vararg args: String) : String{
         return buildString{
